Fixed typo/refactor
