+++ README.md
@@ -20,10 +20,7 @@ It has been bootstrapped using [create-react-app](https://github.com/facebook/cr
 - [x] Formatting of cells (e.g. locale date format, status, etc)
 - [x] Add loading modal when fetching data from `/incidents` endpoint
 - [x] Infinite scrolling using `react-window`
-- [x] Store query/settings using session data
-  - [x] Global Search
-  - [x] Query Settings: Status, Urgency, Priority, Teams, and Services
-  - [x] Table Settings: Column order, sorting, and widths
+- [x] Store settings using session data: Global search, query settings, and table settings.
 
 ##### Code
 
@@ -43,7 +40,7 @@ It has been bootstrapped using [create-react-app](https://github.com/facebook/cr
 - [ ] External system sync (post-action) button (e.g. disabled correctly, but enabled on incident reselect due to incident store not refreshed)
 - [ ] Resolved incidents still appear in the table after a period (most likely memoization of react-table)
 - [x] Updating column order reverts to width defined under `src/util/incident-table-column.js`
-- [ ] Updating columns may occasionally reset back to `minWidth` 
+- [ ] Updating columns may occasionally reset back to `minWidth`
 - [x] `state.querySettings.incidentPriority` resets on render, thereby losing persisted data
 - [ ] `react-select` within query settings does not render existing selections when not in view (needs re-rendering via hook)
