+++ src/redux/store.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware } from 'redux';
 import createSagaMiddleware from 'redux-saga';
 
 import { persistStore, persistReducer } from 'redux-persist';
-import { persistConfig } from './persistance/config';
+import { persistConfig } from './persistence/config';
 
 import rootReducer from './rootReducer';
 import rootSaga from './rootSaga';
